+++ internal/proto_decoder/proto_resolver.go
@@ -11,7 +11,15 @@ type descriptorResolver struct {
 	methodDescriptors map[string]*desc.MethodDescriptor
 }
 
-func (d *descriptorResolver) resolve(fullMethod string, direction internal.MessageOrigin, raw []byte) (*desc.MessageDescriptor, error) {
+func (d *descriptorResolver) resolveEncoded(fullMethod string, message *internal.Message) (*desc.MessageDescriptor, error) {
+	return d.resolve(fullMethod, message.MessageOrigin)
+}
+
+func (d *descriptorResolver) resolveDecoded(fullMethod string, message *internal.Message) (*desc.MessageDescriptor, error) {
+	return d.resolve(fullMethod, message.MessageOrigin)
+}
+
+func (d *descriptorResolver) resolve(fullMethod string, direction internal.MessageOrigin) (*desc.MessageDescriptor, error) {
 	if descriptor, ok := d.methodDescriptors[fullMethod]; ok {
 		switch direction {
 		case internal.ClientMessage: